slackbot/pyproject.toml
[tool.poetry]
name = "slackbot"
version = "0.1.0"
description = ""
authors = ["Sean Mayo <sean.mayo@matsonfruit.com>"]

[tool.poetry.dependencies]
python = "^3.7"
slack-bolt = "^1.9.1"
python-dotenv = "^0.19.0"
pyimgur = "^0.6.0"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
